+++ src/models/IDomainNameEntry.ts
@@ -1,0 +1,5 @@
+export interface IDomainNameEntry {
+	name: string
+	urls: string[]
+	ips: string[]
+}